Fix: switch to image-extensions
