+++ .cjsescache
@@ -1,6 +1,7 @@
 [
   "node_modules/content-disposition/index.js",
   "node_modules/event-lite/event-lite.js",
+  "node_modules/image-extensions/image-extensions.json",
   "node_modules/webext-menus/index.js",
   "node_modules/webext-pref/lib/promisify.js",
   "node_modules/webextension-polyfill/dist/browser-polyfill.js",
