+++ src/background.js
@@ -11,6 +11,7 @@ import {createCounter} from "./lib/counter.js";
 import {IS_CHROME} from "./lib/env.js";
 import {createDialog} from "./lib/popup-dialog.js";
 import {compileStringTemplate} from "./lib/string-template.js";
+import {expandEnv} from "./lib/expand-env.mjs";
 
 const MENU_ACTIONS = {
 	PICK_FROM_CURRENT_TAB: {
@@ -554,52 +555,3 @@ async function singleDownload({url, env, tabId, frameId, referrer, alt}) {
   }, true);
 }
 
-function expandEnv(env) {
-	// image url
-	var url = new URL(env.url);
-	env.hostname = url.hostname;
-	
-	// image filename
-	var base, name, ext;
-	if (env.base) {
-		base = env.base;
-	} else {
-		try {
-			base = url.href.match(/([^/]+)\/?$/)[1];
-		} catch (err) {
-			base = pref.get("defaultName");
-		}
-	}
-	try {
-		[, name, ext] = base.match(/^(.+)(\.(?:jpg|png|gif|jpeg|svg))\b/i);
-	} catch (err) {
-		name = base;
-		ext = pref.get("defaultExt");
-	}
-	env.base = nestDecodeURIComponent(base);
-	env.name = nestDecodeURIComponent(name);
-	env.ext = nestDecodeURIComponent(ext);
-	
-	// page url
-	url = new URL(env.pageUrl);
-	env.pageHostname = url.hostname;
-  env.pageName = pathToName(url.pathname);
-}
-
-function pathToName(path) {
-  const base = path.match(/\/([^/]*)\/?$/)[1];
-  const i = base.lastIndexOf(".");
-  if (i < 0) return base;
-  return base.slice(0, i);
-}
-
-function nestDecodeURIComponent(s) {
-	while (/%[0-9a-f]{2}/i.test(s)) {
-		try {
-			s = decodeURIComponent(s);
-		} catch (err) {
-			break;
-		}
-	}
-	return s;
-}
